Cosmetic: use correct size for comparison of pointers.
By: Brett Nash (mode-mem.patch)


SVN revision: 31699

+++ legacy/evas/src/modules/engines/fb/evas_fb_main.c
@@ -535,7 +535,7 @@ fb_postinit(FB_Mode *mode)
   mode->mem_offset = (unsigned)(fb_fix.smem_start) & (getpagesize()-1);
   mode->mem = (unsigned char *)mmap(NULL, fb_fix.smem_len + mode->mem_offset,
 				 PROT_WRITE | PROT_READ, MAP_SHARED, fb, 0);
-  if ((int)mode->mem == -1)
+  if ((intptr_t)mode->mem == -1)
     {
       perror("mmap");
       fb_cleanup();
